Minor #include and type fixes
Fixes an incorrectly signed integer type and adds a missing include.

PiperOrigin-RevId: 489546971

+++ dpf/distributed_point_function.h
@@ -732,7 +732,7 @@ absl::StatusOr<std::vector<T>> DistributedPointFunction::EvaluateUntil(
     int block_index = DomainToBlockIndex(prefixes[i], previous_hierarchy_level);
 
     // Check if `tree_index` already exists in `tree_indices`.
-    int64_t previous_size = tree_indices_inverse.size();
+    size_t previous_size = tree_indices_inverse.size();
     auto it = tree_indices_inverse.try_emplace(tree_indices_inverse.end(),
                                                tree_index, tree_indices.size());
     if (tree_indices_inverse.size() > previous_size) {
